Move home component inside components folder
