+++ wa-public/src/app/app-routing.module.ts
@@ -4,7 +4,7 @@ import { AcceptDisclaimerComponent } from './components/accept-disclaimer/accept
 import { RequestTokenComponent } from './components/request-token/request-token.component';
 import { KeycloakGuard } from './guards/keycloak.guard';
 import { ValidInviteGuard } from './guards/valid-invite.guard';
-import { HomeComponent } from './home/home.component';
+import { HomeComponent } from './components/home/home.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { CompletedComponent } from './pages/completed/completed.component';
 import { SuccessComponent } from './pages/success/success.component';
